// src/components/gosling.d.ts
export type Track = SingleTrack | OverlaidTrack | TemplateTrack | DummyTrack;
// ...
export interface DummyTrack {
  /**
   * internal
   */
  _invalidTrack?: boolean;
  /**
   * No assemblies can be associated with a dummy track
   */
  assembly?: "unknown";
  /**
   * Specify the track height in pixels.
   */
  height?: number;
  /**
   * Assigned to `uid` in a HiGlass view config, used for API and caching.
   */
  id?: string;
  /**
   * Only linear layout are supported at this time
   */
  layout?: "linear";
  /**
   * Specify the orientation.
   */
  orientation?: "horizontal" | "vertical";
  overlayOnPreviousTrack?: boolean;
  /**
   * Whether to disable [Zooming and Panning](http://gosling-lang.org/docs/user-interaction#zooming-and-panning), __Default:__ `false`.
   */
  static?: true;
  /**
   * Defines how the track is styled
   */
  style?: {
    /**
     * Background color of the track
     */
    background?: string;
    /**
     * Color of the outline of the track
     */
    outline?: string;
    /**
     * Specify the font size of the title
     */
    textFontSize?: number;
    /**
     * Specify the font weight of the title.
     */
    textFontWeight?: "bold" | "normal";
    /**
     * Specify the stroke color of title.
     */
    textStroke?: string;
    /**
     * Specify the stroke width of the title.
     */
    textStrokeWidth?: number;
  };
  /**
   * Text that gets shown on the DummyTrack
   */
  title?: string;
  /**
   * Used to specify the dummy track
   */
  type: "dummy-track";
  /**
   * Specify the track width in pixels.
   */
  width?: number;
  /**
   * Unused property for DummyTrack
   *
   * @minItems 2
   * @maxItems 2
   */
  zoomLimits?: [null, null];
}